requirements.txt
google-genai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0